backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart
requests
PyMuPDF>=1.23.0
pytesseract>=0.3.10
pillow>=10.0.0